Try using true division instead of casting to float

+++ irc/util.py
@@ -1,3 +1,5 @@
+from __future__ import division
+
 # from jaraco.util.itertools
 def always_iterable(item):
     """
@@ -30,5 +32,5 @@ def total_seconds(td):
         result = td.total_seconds()
     except AttributeError:
         seconds = td.seconds + td.days * 24 * 3600
-        result = float((td.microseconds + seconds * 10**6) / 10**6)
+        result = (td.microseconds + seconds * 10**6) / 10**6
     return result
